cleanup

+++ start.sh
@@ -1,14 +1,13 @@
 #!/bin/sh
-echo -e  'y\n'|ssh-keygen -q -t ecdsa -N "" -f ~/.ssh/id_ecdsa > /dev/null 2>&1
-rm  -f /root/.ssh/authorized_keys
-#curl $KEY_URL |xargs -n1 -I{} echo '{}'>> /tmp/rootsshauthorized_keys
-wget $KEY_URL -O /tmp/rootsshauthorized_keys
-cat /tmp/rootsshauthorized_keys|sort|uniq>/tmp/rootsshauthorized_keysuniq
-rm /tmp/rootsshauthorized_keys
-mv /tmp/rootsshauthorized_keysuniq /root/.ssh/authorized_keys
 
-chmod 600 /root/.ssh/authorized_keys
+
+if [ ! -f /root/.ssh/id_ecdsa.pub ]
+  then
+    echo -e  'y\n'|ssh-keygen -q -t ecdsa -N "" -f ~/.ssh/id_ecdsa > /dev/null 2>&1
+fi
+
 ssh-keygen -A
+curl $KEY_URL|sort|uniq>/root/.ssh/authorized_keys
+chmod 600 /root/.ssh/authorized_keys
 
-echo 'starting sshd'
 exec /usr/sbin/sshd -D -e "$@"
